requirements (2).txt
streamlit
requests
newspaper3k
groq
lxml[html_clean]
diffusers
torch
transformers
IPython
